requirements_xtts.txt
coqui-tts>=0.27.2
trainer>=0.0.36
coqpit>=0.0.17
huggingface_hub>=0.25.0
datasets>=2.18.0
librosa>=0.10.2
soundfile>=0.12.1
numpy<2
scipy>=1.11.0
tqdm>=4.66.0
